# docs/swagger.yaml
definitions:
  gin.H:
    additionalProperties: {}
    type: object
...
  github_com_EvolutionAPI_evolution-go_pkg_poll_model.PollResults:
    properties:
      optionCounts:
        additionalProperties:
          type: integer
        description: hash -> count
        type: object
      pollChatJid:
        type: string
      pollMessageId:
        type: string
      totalVotes:
        type: integer
      voters:
        items:
          $ref: '#/definitions/github_com_EvolutionAPI_evolution-go_pkg_poll_model.VoterInfo'
        type: array
      votes:
        items:
          $ref: '#/definitions/github_com_EvolutionAPI_evolution-go_pkg_poll_model.PollVote'
        type: array
    type: object
  github_com_EvolutionAPI_evolution-go_pkg_poll_model.PollVote:
    properties:
      companyId:
        type: string
      id:
        type: string
      instanceId:
        type: string
      pollChatJid:
        type: string
      pollMessageId:
        type: string
      receivedAt:
        type: string
      selectedOptions:
        description: SHA-256 hashes
        items:
          type: string
        type: array
      voteMessageId:
        type: string
      votedAt:
        type: string
      voterJid:
        type: string
      voterName:
        type: string
      voterPhone:
        type: string
    type: object
  github_com_EvolutionAPI_evolution-go_pkg_poll_model.VoterInfo:
    properties:
      jid:
        type: string
      name:
        type: string
      phone:
        type: string
      selectedOptions:
        items:
          type: string
        type: array
      votedAt:
        type: string
    type: object
...
info:
  contact: {}
  description: Evolution GO - whatsmeow
  title: Evolution GO
  version: "1.0"
paths:
...
  /polls/{pollMessageId}/results:
    get:
      consumes:
      - application/json
      description: Retorna todos os votos de uma enquete específica
      parameters:
      - description: ID da mensagem da enquete
        in: path
        name: pollMessageId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_EvolutionAPI_evolution-go_pkg_poll_model.PollResults'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Get poll results
      tags:
      - Polls
...
swagger: "2.0"